requirements.txt
streamlit
osmnx
geopandas
folium
streamlit-folium
ultralytics
requests
opencv-python-headless
Pillow
shapely
numpy
pandas
